adjusted step by step for recordings downloader, changes saved in recordings-downloader folder
